more linking fixes
