+++ test/gtest/conv_f8_bwd.cpp
@@ -29,11 +29,11 @@
 #include <serialize.hpp>
 #include "tensor_util.hpp"
 #include "get_handle.hpp"
-#include "conv3d_test_case.hpp"
 #include "f8_cast_util.hpp"
+#include "conv3d_test_case.hpp"
 
+namespace conv_f8_bwd {
 
-template<>
 std::vector<Conv3DTestCase> ConvTestConfigs()
 { // g    n   c   d    h   w   k   z  y  x pad_x pad_y pad_z stri_x stri_y stri_z dia_x dia_y dia_z
     return {{1, 16, 16, 1, 14, 14, 16, 1, 3, 3, 1, 1, 0, 1, 1, 1, 1, 1, 1, miopenConvolution},
@@ -45,7 +45,6 @@ std::vector<Conv3DTestCase> ConvTestConfigs()
             {4, 128, 4, 1, 28, 28, 4, 1, 3, 3, 1, 1, 0, 1, 1, 1, 1, 1, 1, miopenConvolution},
             {2, 128, 2, 1, 28, 28, 2, 1, 3, 3, 1, 1, 0, 1, 1, 1, 1, 1, 1, miopenConvolution}};
 }
-namespace conv_f8_bwd {
 
 template <typename T = float>
 struct ConvBwdSolverTest
@@ -193,7 +192,7 @@ TEST_P(ConvBwdSolverTestF8, CKConvF8Bwd)
 INSTANTIATE_TEST_SUITE_P(ConvBwdTest,
                          ConvBwdSolverTestF8,
                          testing::Combine(testing::Values(miopenConvolutionBwdDataAlgoImplicitGEMM),
-                                          testing::ValuesIn(ConvTestConfigs<Conv3DTestCase>()),
+                                          testing::ValuesIn(ConvTestConfigs()),
                                           testing::Values(miopenTensorNDHWC)));
 
 } //namespace conv_f8_bwd
