+++ test/gtest/group_conv3d_wrw.cpp
@@ -32,6 +32,21 @@
 #include "get_handle.hpp"
 #include "group_conv3d_wrw.hpp"
 
+namespace group_conv3d_wrw {
+
+std::vector<Conv3DTestCase> ConvTestConfigs()
+{ // g   n   c   d    h   w   k   z  y  x pad_x pad_y pad_z stri_x stri_y stri_z dia_x dia_y dia_z
+    return {{1, 128, 64, 14, 28, 28, 64, 3, 3, 3, 1, 1, 1, 1, 1, 1, 1, 1, 1, miopenConvolution},
+            {1, 64, 32, 28, 28, 28, 32, 3, 3, 3, 1, 1, 1, 1, 1, 1, 1, 1, 1, miopenConvolution},
+            {2, 128, 32, 28, 28, 28, 32, 3, 3, 3, 1, 1, 1, 1, 1, 1, 1, 1, 1, miopenConvolution},
+            {32, 128, 32, 28, 28, 28, 32, 3, 3, 3, 1, 1, 1, 1, 1, 1, 1, 1, 1, miopenConvolution},
+            {8, 64, 32, 28, 28, 28, 32, 3, 3, 3, 1, 1, 1, 1, 1, 1, 1, 1, 1, miopenConvolution},
+            {16, 64, 32, 28, 28, 28, 32, 3, 3, 3, 1, 1, 1, 1, 1, 1, 1, 1, 1, miopenConvolution},
+            {3, 48, 48, 28, 28, 28, 48, 3, 3, 3, 1, 1, 1, 1, 1, 1, 1, 1, 1, miopenConvolution},
+            {3, 48, 39, 28, 28, 28, 39, 3, 3, 3, 1, 1, 1, 1, 1, 1, 1, 1, 1, miopenConvolution},
+            {5, 120, 60, 28, 28, 28, 60, 3, 3, 3, 1, 1, 1, 1, 1, 1, 1, 1, 1, miopenConvolution}};
+}
+
 struct ConvWrwSolverTest3D : ConvWrwSolverTest<float>
 {
 };
@@ -95,5 +110,7 @@ INSTANTIATE_TEST_SUITE_P(
     ConvWrwTest,
     ConvWrwSolverTest3D,
     testing::Combine(testing::Values(miopenConvolutionBwdWeightsAlgoImplicitGEMM),
-                     testing::ValuesIn(ConvTestConfigs<Conv3DTestCase>()),
+                     testing::ValuesIn(ConvTestConfigs()),
                      testing::Values(miopenTensorNDHWC)));
+
+} //namespace group_conv3d_wrw
